// src/app/pages/items/models/assessment-filter-information.ts
export class AssessmentFilterInformation {
  itemType: string = null;
  tagIds: string[] = [];
  totalQuestions: number = null;
  subjectId: string = null;
  subjectName?: string = null; 
  topicId: string = null;
  topicName: string = null;
  sectionId: string = null;
  blockId: number = null;
  content: string = null;
  subtopicId: string = null;
  subtopicName: string = null;
  score: number;
  difficultyLevel: number;
}